requirements.txt
# API and Web Framework
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# OpenAI Integration
openai==1.12.0
aiohttp==3.9.3

# Data Processing
python-dateutil==2.8.2

# Environment and Configuration
python-dotenv==1.0.1

# Database
pymongo==4.6.0

# HTTP Utilities
requests==2.31.0

# Development Tools
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0 